requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.29
gunicorn==20.1.0
Werkzeug==2.0.3
google-cloud-storage
